Updated various libraries, updated build.gradle to match, update cache code for changes in ehcache 2.7

+++ framework/src/api/java/org/moqui/context/Cache.java
@@ -30,9 +30,9 @@ public interface Cache {
     long getExpireTimeLive();
     void setExpireTimeLive(long expireTime);
 
-    int getMaxElements();
+    long getMaxElements();
     EvictionStrategy getEvictionStrategy();
-    void setMaxElements(int maxSize, EvictionStrategy strategy);
+    void setMaxElements(long maxSize, EvictionStrategy strategy);
 
     Object get(Serializable key);
     Object put(Serializable key, Object value);
